+++ app/src/main/java/ru/practicum/android/diploma/data/db/converters/ConverterIntoModel.kt
@@ -1,5 +1,9 @@
-package ru.practicum.android.diploma.data.db.entity
+package ru.practicum.android.diploma.data.db.converters
 
+import ru.practicum.android.diploma.data.db.entity.AreaEntity
+import ru.practicum.android.diploma.data.db.entity.KeySkillEntity
+import ru.practicum.android.diploma.data.db.entity.PhoneEntity
+import ru.practicum.android.diploma.data.db.entity.VacancyEntity
 import ru.practicum.android.diploma.domain.models.Vacancy
 import ru.practicum.android.diploma.domain.models.VacancyDetails
 import ru.practicum.android.diploma.domain.models.components.AreaModel
@@ -13,83 +17,7 @@ import ru.practicum.android.diploma.domain.models.components.PhonesModel
 import ru.practicum.android.diploma.domain.models.components.SalaryModel
 import ru.practicum.android.diploma.domain.models.components.ScheduleModel
 
-class ConverterDb() {
-    fun intoVacancyEntity(vacancy: VacancyDetails): VacancyEntity {
-        return VacancyEntity(
-            id = vacancy.id,
-            idAreaModel = vacancy.area?.id,
-            nameAreaModel = vacancy.area?.name,
-            countryId = vacancy.area?.countryId,
-            idEmploymentModel = vacancy.employment?.id,
-            original = vacancy.employer?.logoUrls?.original,
-            logo90 = vacancy.employer?.logoUrls?.logo90,
-            logo240 = vacancy.employer?.logoUrls?.logo240,
-            nameEmployerModel = vacancy.employer?.name,
-            trusted = vacancy.employer?.trusted,
-            url = vacancy.employer?.url,
-            vacanciesUrl = vacancy.employer?.vacanciesUrl,
-            name = vacancy.name,
-            currency = vacancy.salary?.currency,
-            from = vacancy.salary?.from,
-            gross = vacancy.salary?.gross,
-            to = vacancy.salary?.to,
-            description = vacancy.description,
-            idEmployerModel = vacancy.employment?.id,
-            nameEmploymentModel = vacancy.employment?.name,
-            idExperienceModel = vacancy.experience?.id,
-            nameExperienceModel = vacancy.experience?.name,
-            email = vacancy.contacts?.email,
-            nameContactsModel = vacancy.contacts?.name,
-            idScheduleModel = vacancy.schedule?.id,
-            nameScheduleModel = vacancy.schedule?.name,
-            alternateUrl = vacancy.alternateUrl
-        )
-    }
-
-    fun intoPhoneEntity(vacancy: VacancyDetails): List<PhoneEntity> {
-        val listPhone = ArrayList<PhoneEntity>()
-        vacancy.contacts?.phones?.forEach { phonesModel ->
-            listPhone.add(
-                PhoneEntity(
-                    idVacancy = vacancy.id.toInt(),
-                    cityCode = phonesModel.cityCode,
-                    comment = phonesModel.comment,
-                    countryCode = phonesModel.countryCode,
-                    formatted = phonesModel.formatted,
-                    number = phonesModel.number
-                )
-            )
-        }
-        return listPhone
-    }
-
-    fun intoKeySkillEntity(vacancy: VacancyDetails): List<KeySkillEntity> {
-        val listKey = ArrayList<KeySkillEntity>()
-        vacancy.keySkills?.forEach { keySkillsModel ->
-            listKey.add(
-                KeySkillEntity(
-                    idVacancy = vacancy.id.toInt(),
-                    name = keySkillsModel.name
-                )
-            )
-        }
-        return listKey
-    }
-
-    fun intoAreaEntity(vacancy: VacancyDetails): List<AreaEntity> {
-        val listArea = ArrayList<AreaEntity>()
-        vacancy.area?.areas?.forEach { area ->
-            listArea.add(
-                AreaEntity(
-                    idVacancy = vacancy.id.toInt(),
-                    idArea = area.id,
-                    name = area.name,
-                    countryId = area.countryId
-                )
-            )
-        }
-        return listArea
-    }
+class ConverterIntoModel {
 
     fun intoPhone(phones: List<PhoneEntity>): List<PhonesModel> {
         val listPhone = ArrayList<PhonesModel>()
@@ -137,7 +65,7 @@ class ConverterDb() {
         areas: List<AreaEntity>
     ): Vacancy {
         return Vacancy(
-            id = vacancy.id,
+            id = vacancy.id.toString(),
             area = AreaModel(
                 id = vacancy.idAreaModel,
                 name = vacancy.nameAreaModel,
@@ -174,25 +102,14 @@ class ConverterDb() {
         areaEntity: List<AreaEntity>
     ): VacancyDetails {
         return VacancyDetails(
-            id = vacancyEntity.id,
+            id = vacancyEntity.id.toString(),
             area = AreaModel(
                 id = vacancyEntity.idAreaModel,
                 name = vacancyEntity.nameAreaModel,
                 countryId = vacancyEntity.countryId,
                 areas = intoArea(areaEntity)
             ),
-            employer = EmployerModel(
-                id = vacancyEntity.idEmployerModel,
-                logoUrls = LogoUrlModel(
-                    original = vacancyEntity.original,
-                    logo90 = vacancyEntity.logo90,
-                    logo240 = vacancyEntity.logo240,
-                ),
-                name = vacancyEntity.nameEmployerModel,
-                trusted = vacancyEntity.trusted,
-                url = vacancyEntity.url,
-                vacanciesUrl = vacancyEntity.vacanciesUrl
-            ),
+            employer = getEmployerModel(vacancyEntity),
             name = vacancyEntity.name,
             salary = SalaryModel(
                 currency = vacancyEntity.currency,
@@ -221,6 +138,22 @@ class ConverterDb() {
             keySkills = intoKeySkill(listKey),
             alternateUrl = vacancyEntity.alternateUrl
         )
+    }
 
+    private fun getEmployerModel(
+        vacancyEntity: VacancyEntity
+    ): EmployerModel {
+        return EmployerModel(
+            id = vacancyEntity.idEmployerModel,
+            logoUrls = LogoUrlModel(
+                original = vacancyEntity.original,
+                logo90 = vacancyEntity.logo90,
+                logo240 = vacancyEntity.logo240,
+            ),
+            name = vacancyEntity.nameEmployerModel,
+            trusted = vacancyEntity.trusted,
+            url = vacancyEntity.url,
+            vacanciesUrl = vacancyEntity.vacanciesUrl
+        )
     }
 }
