+++ app/src/main/java/ru/practicum/android/diploma/di/RepositoryModule.kt
@@ -4,10 +4,15 @@ import org.koin.dsl.module
 import ru.practicum.android.diploma.data.db.SelectedVacanciesRepositoryImpl
 import ru.practicum.android.diploma.data.impl.SearchRepositoryImpl
 import ru.practicum.android.diploma.domain.api.SearchRepository
+import ru.practicum.android.diploma.domain.db.SelectedVacanciesRepository
 
 val repositoryModule = module {
-    single<SelectedVacanciesRepositoryImpl> {
-        SelectedVacanciesRepositoryImpl(get(),get())
+    single<SelectedVacanciesRepository> {
+        SelectedVacanciesRepositoryImpl(
+            vacanciesDatabase = get(),
+            converterModel = get(),
+            converterIntoEntity = get()
+        )
     }
 
     single<SearchRepository> {
